Moved zpik and zunpik in general
